// src/lib/prisma/generated/models/SidebarMenuItem.ts
export type SidebarMenuItemCreateInput = {
  id: number
  icon: string
  text: string
  destination: string
  displayOrder: number
  createdAt?: Date | string
  updatedAt?: Date | string
}